+++ src/assets/Styles/LoginRegister.css
@@ -1,182 +1,260 @@
 @import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,600;0,700;1,500&display=swap');
 
-
-.Login-Register{
-    display: flex;
-    justify-content: space-around;
-    padding: 40px 0px 0px 80px;
+.form-container form {
+	background-color: #FFFFFF;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	flex-direction: column;
+	padding: 0 50px;
+	height: 100%;
 }
 
-.logo-sing-in{
-    display: flex;
-    margin-left: -140px;
-    margin-bottom: 70px;
+.form-container {
+	position: absolute;
+	top: 0;
+	height: 100%;
+	transition: all .6s ease-in-out;
 }
 
-.name-empresa{
-    color: #3CC0C9;
-    font-family: Poppins, Bold;
-    padding-left: 15px;
-    font-size: 30px;
-}
-.form-login h2 {
-    color: #3CC0C9;
-    font-family: Poppins, Bold;
-    font-size: 40px;
-    margin: 0;
+.sign-in-container {
+	left: 0px;
+	width: 50%;
+	z-index: 2;
 }
 
-.logo-sing-in h2{
-    font-size: 30px;
+.container.right-panel-active .sign-in-container {
+	transform: translateX(100%);
 }
 
-.form-login p{
-    color: #959595;
-    font-family: Poppins, Medium;
+.sign-up-container {
+	left: 0;
+	width: 50%;
+	opacity: 0;
+	z-index: 1;
 }
 
-.bienvenida{
-    margin-left: -18px;
+.container.right-panel-active .sign-up-container {
+	transform: translateX(100%);
+	opacity: 1;
+	z-index: 5;
+	animation: show 0.6s;
 }
 
-.Datos-Form{
-    display: flex;
-    flex-direction: column;
-    padding-top: 30px;
-    margin-left: -55px;
+.overlay-container {
+	position: absolute;
+	top: 0;
+	left: 50%;
+	width: 50%;
+	height: 100%;
+	overflow: hidden;
+	transition: transform 0.6s ease-in-out;
+	z-index: 100;
 }
 
-.email{
-    padding-bottom: 15px;
+.container.right-panel-active .overlay-container {
+	transform: translateX(-100%);
 }
 
-.recuerdame{
-    display: flex;
+.overlay {
+	background: #FFFFFF;
+	color: #5F4A33;
+	position: relative;
+	left: -100%;
+	height: 100%;
+	width: 200%;
+	transform: translateX(0);
+	transition: transform 0.6s ease-in-out;
 }
 
-.recuerdame p{
-    color: #464646;
-    padding-left: 5px;
+.container.right-panel-active .overlay {
+	transform: translateX(50%);
 }
 
-.btn-login button{
-    width: 300px;
-    height: 50px;
-    background-color: #3CC0C9;
-    border: none;
-    border-radius: 10px;
-    font-family: Poppins, Medium;
-    color: #F8FDFD;
-    font-size: 20px;
+.overlay-panel {
+	position: absolute;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	flex-direction: column;
+	padding: 0 40px;
+	text-align: center;
+	top: 0;
+	height: 100%;
+	width: 50%;
+	transform: translateX(0);
+	transition: transform 0.6s ease-in-out;
 }
 
-.btn-login{
-    margin-left: 50px;
+.overlay-left {
+	transform: translateX(-20%);
+	background-image: url('../Images/ImgRegister.svg');
 }
 
-.singUp{
-    display: flex;
-    margin-left: -18px;
+.container.right-panel-active .overlay-left {
+	transform: translateX(0);
 }
 
-.hidden{
-    visibility: hidden;
+.overlay-right {
+	right: 0;
+	transform: translateX(0);
+	background-image: url('../Images/ImgLogin.svg');
+	margin-right: -40px;
 }
 
-/*REGISTER*/
-
-.form-register{
-    margin-left: 120px;
+.container.right-panel-active .overlay-right {
+	transform: translateX(20%);
 }
 
-.form-register h2 {
-    color: #3CC0C9;
-    font-family: Poppins, Bold;
-    font-size: 40px;
-    margin: 0;
-}
-
-.logo-sing-up{
-    display: flex;
-    margin-left: -140px;
-    margin-bottom: 70px;
-}
-
-.logo-sing-up h2{
-    font-size: 30px;
+.reg-fo{
+	font-family: Poppins;
+	color: #3CC0C9;
+	font-size: 35px;
+	margin: 0;
 }
 
 .saludo{
-    margin-left: -25px;
+	padding-bottom: 15px;
 }
 
 .saludo p{
-    font-family: Poppins, Bold;
-    color: #959595;
+	font-family: Poppins;
+	font-size: 18px;
+	color: #959595;
 }
 
 .User-Telefono{
-    display: flex;
-    padding-top: 30px;
-    margin-left: -55px;
-}
-
-.tel{
-    padding-left: 15px;
-}
-
-.E-mail{
-    padding-top: 15px;
-    margin-left: -55px;
+	display: flex;
+	justify-content: center;
+	padding-bottom: 20px;
 }
 
 .contras{
-    display: flex;
-    padding-top: 15px;
-    margin-left: -55px;
+	display: flex;
+	justify-content: center;
+	padding-top: 20px;
+	padding-bottom: 20px;
 }
 
 .password{
-    padding-right: 15px;
+	padding-right: 20px;
+	
 }
 
-.btn-registrarse button{
-    width: 300px;
-    height: 50px;
-    background-color: #3CC0C9;
-    border: none;
-    border-radius: 10px;
-    font-family: Poppins, Medium;
-    color: #F8FDFD;
-    font-size: 20px;
+.tel{
+	padding-left: 15px;
 }
 
-.btn-registrarse{
-    padding-top: 30px;
-    margin-left: 20px;
+.btn-registro{
+	width: 300px;
+	height: 50px;
+	border: none;
+	border-radius: 15px;
+	background-color: #3CC0C9;
+	color: #FFFFFF;
+	font-family: Poppins, Medium;
+	font-size: 20px;
 }
 
-.inicioSesion{
-    margin-left: 20px;
+.btn-inicio-ses{
+	padding-top: 5px;
 }
 
-.inicioSesion p{
-    font-family: Poppins;
-    color: #A5A5A5;
+.btn-inicio-ses button{
+	border: none;
+	background-color: white;
+	font-weight: Bold;
+	font-size: 18px;
+	font-family: Poppins;
+	color: #464646;
 }
 
-.inicioSesion p button{
-    border: none;
-    background: none;
-    font-family: Poppins;
+.btn-inicio-ses p {
+	font-family: Poppins;
+	color: #A5A5A5;
 }
 
-.singUp p{
-    font-family: Poppins;
+.logo-sing-in{
+	display: flex;
+	margin-left: -385px;
+	padding-bottom: 60px;
+	margin-top: -100px;
 }
 
-.singUp p button{
-    border: none;
-    background: none;
-    font-family: Poppins;
+.logo-sing-in h2{
+	font-size: 35px;
+}
+
+.logo-sing h2{
+	font-size: 35px;
+}
+
+.logo-sing{
+	display: flex;
+	margin-left: -375px;
+	padding-bottom: 60px;
+	margin-top: -100px;
+}
+
+.recuerdame{
+	display: flex;
+	padding-top: 15px;
+	margin-left: -270px;
+}
+
+.recuerdame p{
+	padding-left: 10px;
+	color: #464646;
+}
+
+.name-empresa{
+	color: #3CC0C9;
+	font-family: Poppins;
+	padding-left: 15px;
+	font-size: 30px;
+}
+
+.inicioS{
+	color: #3CC0C9;
+	font-family: Poppins;
+	font-size: 40px;
+	margin: 0;
+}
+
+.con-login{
+	color: #959595;
+	font-family: Poppins;
+}
+
+.email{
+	padding-top: 25px;
+	padding-bottom: 20px;
+}
+
+.btn-login{
+	padding-top: 0px;
+}
+
+.btn-login button{
+	width: 300px;
+	height: 50px;
+	border: none;
+	border-radius: 15px;
+	background-color: #3CC0C9;
+	color: #FFFFFF;
+	font-family: Poppins, Medium;
+	font-size: 20px;
+}
+
+.iniciar-registro{
+	padding-top: 10px;
+}
+
+.iniciar-registro button{
+	border: none;
+	background-color: white;
+	font-weight: Bold;
+	font-size: 18px;
+	font-family: Poppins;
+	color: #464646;
 }
